WIP: Initial verison of the script editor
Store entity-scripts in the asset server
