+++ scripts/system/create/entityProperties/html/entityProperties.html
@@ -57,8 +57,6 @@
                     <td class="tabsPropertiesFrame">
                         <div class="tabsPropertiesPage" id="properties-pages">
                             <!-- each property is added at runtime in entityProperties -->
-<!-- div class="labelTabHeader">SCRIPTS</div></div><div id="div-property-script" class="property container"><label><span class="" disabled="disabled">Script22</span></label><div style="width: 100%;" class="url refresh"><input id="property-script" type="text" placeholder="URL" disabled="disabled"><span class="multi-diff" disabled="disabled"></span><input class="glyph" type="button" id="property-script-button-editor" value="s" disabled="disabled"><input class="glyph" type="button" id="property-script-button-reload" value="F" disabled="disabled"></div></div><div id="div-property-serverScripts" class="property container"><label><span class="" disabled="disabled">Server Script22</span></label><div style="width: 100%;" class="url refresh"><input id="property-serverScripts" type="text" placeholder="URL" disabled="disabled"><span class="multi-diff" disabled="disabled"></span><input class="glyph" type="button" id="property-serverScripts-button-editor" value="s" disabled="disabled"><input class="glyph" type="button" id="property-serverScripts-button-reload" value="F" disabled="disabled"></div></div><div id="div-property-serverScriptStatus" class="property container"><label><span class="indented" disabled="disabled">Server Script Status</span></label><div style="width: 100%;"><div id="property-serverScripts-status">Not running</div><div class="property" style="display: none;"><textarea id="property-serverScripts-error" scrolling="false" disabled="disabled"></textarea></div></div></div><div id="div-property-userData" class="property container"><label><span class="" disabled="disabled">User Data</span></label><div style="width: 100%;" class="textarea"><div id="property-userData-editor"></div><div id="property-userData-editorStatus"></div><textarea id="property-userData" scrolling="false" disabled="disabled"></textarea><span class="multi-diff" disabled="disabled"></span><br><div class="row"><input class="red" type="button" id="property-userData-button-clear" value="Clear User 
-Data" disabled="disabled"><input class="blue" type="button" id="property-userData-button-edit" value="Edit as JSON" disabled="disabled"><input class="black" type="button" id="property-userData-button-save" value="Save User Data" disabled="disabled"></div></div></div></div><div class="section major" id="properties-collision" style="display: none;"><div class="tab-section-header" -->
                         </div>
                     </td>
                 </tr>
